requirements.txt
astropy==4.2
corner==2.1.0
healpy==1.14.0
ligo.skymap==0.5.0
matplotlib==3.3.4
matplotlib==3.3.4
pandas==1.2.2
seaborn==0.11.1
tables==3.6.1
